refactor(backend): migrate server changes
- eslint autofix
- move `server` folder into `src`
- remove `use strict`
- migrate devDependencies to dependencies
