+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.0",
   "description": "creamy-monitor simplifies monitoring of network-connected services.",
   "scripts": {
-    "start": "node server/server.js",
+    "start": "node src/server",
     "lint": "eslint src test",
     "cover": "nyc --reporter=lcov --reporter=text-summary npm test",
     "knex": "knex --knexfile=src/database/knexfile.js",
@@ -13,11 +13,17 @@
   "author": "CreamyThings <cream@albinodrought.com>",
   "license": "AGPL-3.0",
   "dependencies": {
+    "cors": "^2.8.4",
     "dotenv": "^6.0.0",
+    "express": "^4.16.3",
+    "jsonwebtoken": "^8.3.0",
     "knex": "^0.15.2",
+    "morgan": "^1.9.1",
     "mysql2": "^1.6.1",
     "objection": "^1.3.0",
+    "passport": "^0.4.0",
     "passport-github": "^1.1.0",
+    "passport-jwt": "^4.0.0",
     "request": "^2.88.0"
   },
   "devDependencies": {
@@ -26,19 +32,11 @@
     "@commitlint/travis-cli": "^7.1.2",
     "body-parser": "^1.18.3",
     "cookie-parser": "^1.4.3",
-    "cors": "^2.8.4",
     "eslint": "^5.6.1",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.14.0",
-    "evil-dns": "^0.2.0",
-    "express": "^4.16.3",
     "husky": "^1.1.1",
-    "jest": "^23.6.0",
-    "jsonwebtoken": "^8.2.0",
-    "morgan": "^1.9.0",
-    "passport": "^0.4.0",
-    "passport-jwt": "^3.0.1",
-    "passport-local": "^1.0.0"
+    "jest": "^23.6.0"
   },
   "husky": {
     "hooks": {
