--- server/README.md
@@ -1,2 +1,0 @@
-# Creamy Moniter Server
-
