--- server/routes/test.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const express = require('express');
-const router = express.Router();
-
-// Get request test
-router.get('/test', (req, res, next) => {
-  res.json({ message: 'Working!' });
-});
-
-module.exports = router;
