+++ src/passport/jwt.js
@@ -1,13 +1,11 @@
-'use strict';
-
 const { Strategy: JwtStrategy, ExtractJwt } = require('passport-jwt');
 
-const { JWT_SECRET } = require('../config');
+const { JWT_SECRET } = require('../config/server');
 
 const options = {
   secretOrKey: JWT_SECRET,
   jwtFromRequest: ExtractJwt.fromAuthHeaderWithScheme('Bearer'),
-  algorithms: ['HS256']
+  algorithms: ['HS256'],
 };
 
 const jwtStrategy = new JwtStrategy(options, (payload, done) => {
