+++ src/routes/authTest.js
@@ -1,10 +1,9 @@
-'use strict';
-
 const express = require('express');
+
 const router = express.Router();
 
 // Protected GET
-router.get('/protected', (req, res, next) => {
+router.get('/protected', (req, res) => {
   res.json({ message: 'Protected Endpoint' });
 });
