+++ src/server.js
@@ -1,5 +1,3 @@
-'use strict';
-
 // allow dotenv to populate
 require('dotenv').config();
 
@@ -8,18 +6,16 @@ const morgan = require('morgan');
 const cors = require('cors');
 const { Model } = require('objection');
 const passport = require('passport');
-const evilDns = require('evil-dns');
 
 const knex = require('../src/database/bootstrap');
 
 const githubStrategy = require('./passport/github');
 const jwtStrategy = require('./passport/jwt');
 
-const testRouter = require('./routes/test');
 const authTestRouter = require('./routes/authTest');
 const authRouter = require('./routes/auth');
 
-const { PORT, CLIENT_ORIGIN } = require('./config');
+const { PORT, CLIENT_ORIGIN } = require('./config/server');
 
 const app = express();
 
@@ -30,15 +26,15 @@ passport.use(jwtStrategy);
 // Log all requests. Skip logging during
 app.use(
   morgan(process.env.NODE_ENV === 'development' ? 'dev' : 'common', {
-    skip: () => process.env.NODE_ENV === 'test'
-  })
+    skip: () => process.env.NODE_ENV === 'test',
+  }),
 );
 
 // CORS ACCESS
 app.use(
   cors({
-    origin: CLIENT_ORIGIN
-  })
+    origin: CLIENT_ORIGIN,
+  }),
 );
 
 // Create a static webserver
@@ -50,7 +46,6 @@ app.use(express.json());
 // required for passport to work
 app.use(passport.initialize());
 
-app.use('/api', testRouter);
 app.use('/api/auth', authRouter);
 
 // Endpoints below this require a valid JWT
@@ -59,7 +54,7 @@ app.use(passport.authenticate('jwt', { session: false, failWithError: true }));
 app.use('/api', authTestRouter);
 
 // Catch-all 404
-app.use(function(req, res, next) {
+app.use((req, res, next) => {
   const err = new Error('Not Found');
   err.status = 404;
   next(err);
@@ -67,12 +62,12 @@ app.use(function(req, res, next) {
 
 // Catch-all Error handler
 // Add NODE_ENV check to prevent stacktrace leak
-app.use(function(err, req, res, next) {
+app.use((err, req, res) => {
   console.log(err.message);
   res.status(err.status || 500);
   res.json({
     message: err.message,
-    error: app.get('env') === 'development' ? err : {}
+    error: app.get('env') === 'development' ? err : {},
   });
 });
 
@@ -80,19 +75,14 @@ app.use(function(err, req, res, next) {
 if (require.main === module) {
   // init knex and pass off to objection
   Model.knex(
-    knex()
+    knex(),
   );
 
-  if (process.env.NODE_ENV === 'development') {
-    // force back to localhost for development
-    evilDns.add('albinodrought.com', '127.0.0.1')
-  }
-
   app
-    .listen(PORT, function() {
+    .listen(PORT, function () {
       console.info(`Server listening on ${this.address().port}`);
     })
-    .on('error', err => {
+    .on('error', (err) => {
       console.error(err);
     });
 }
